refactor: move OpenShift connect task runner logic to inspect task runner
Relates to JIRA: DISCOVERY-784
